requirements.txt
Django>=5.2.3
Pillow>=11.2.1
moviepy>=2.2.1
